refactor: move dense traj to package
