+++ oxspires_tools/lidar_undistortion/gtsam_opt.py
@@ -8,7 +8,7 @@ import pandas as pd
 import yaml
 from tqdm import tqdm
 
-from oxspires_tools.lidar_undistortion.core import derive_initial_state
+from oxspires_tools.lidar_undistortion.core import derive_initial_state, integrate_imu
 from oxspires_tools.se3 import se3_matrix_to_xyz_quat_xyzw
 from oxspires_tools.sensor import Sensor
 from oxspires_tools.trajectory.file_interfaces.timestamp import TimeStamp
@@ -330,3 +330,34 @@ def compare_with_vilens(
     )
     print(f"{'bias_acc RMS':<20} {rms(ba_diffs)[0]:>10.5f} {rms(ba_diffs)[1]:>10.5f} {rms(ba_diffs)[2]:>10.5f}")
     print(f"{'bias_gyr RMS':<20} {rms(bg_diffs)[0]:>10.6f} {rms(bg_diffs)[1]:>10.6f} {rms(bg_diffs)[2]:>10.6f}")
+
+
+def build_dense_trajectory(gt_states: list, imu_df: pd.DataFrame, T_BI_mat: np.ndarray):
+    """Integrate IMU segment-by-segment between GT keyframes to produce a dense T_WB trajectory."""
+    T_IB = np.linalg.inv(T_BI_mat)
+    dense_ts = []
+    dense_poses_WB = []
+    for i in range(len(gt_states) - 1):
+        ts_ns_i, T_WB_i, vel_i, bias_acc_i, bias_gyr_i = gt_states[i]
+        ts_ns_j = gt_states[i + 1][0]
+        imu_window = imu_df[(imu_df["timestamp_ns"] >= ts_ns_i) & (imu_df["timestamp_ns"] <= ts_ns_j)].reset_index(
+            drop=True
+        )
+        if len(imu_window) < 2:
+            continue
+        timestamps_ns, poses_WI = integrate_imu(
+            imu_window,
+            initial_velocity=vel_i,
+            initial_pose_WI=T_WB_i @ T_BI_mat,
+            bias_acc=bias_acc_i,
+            bias_gyr=bias_gyr_i,
+        )
+        for t, T_WI in zip(timestamps_ns, poses_WI):
+            if dense_ts and t == dense_ts[-1]:
+                print(
+                    f"Warning: {t} in dense trajectory is exactly on a GT keyframe timestamp; skipping to avoid duplicate"
+                )
+                continue
+            dense_ts.append(t)
+            dense_poses_WB.append(T_WI @ T_IB)
+    return dense_ts, dense_poses_WB
